+++ changelog.md
@@ -1,3 +1,8 @@
+# Ataxx.cpp changelog
+
+Runner used: https://github.com/kz04px/cuteataxx
+Book used: https://github.com/kz04px/cuteataxx/blob/master/res/openings/ply3_fair.txt
+
 ## Version 1
 
 ### 1.0
@@ -6,4 +11,24 @@
 
 ```
 bestmove f1
+```
+
+### 1.1
+
+Material based evaluation, brute-force search, simple timer for stopping.
+
+```
+info depth 1 time 1 score 1 nodes 16 nps 16000 pv f1
+info depth 2 time 1 score 0 nodes 288 nps 288000 pv f1 b1
+info depth 3 time 1 score 1 nodes 7020 nps 7020000 pv f1 b1 e1
+info depth 4 time 2 score 0 nodes 169640 nps 84820000 pv f1 b1 f2 c1
+info depth 5 time 59 score 1 nodes 5084928 nps 86185220 pv f1 a2 e1 a3 d1
+info depth 6 time 1849 score -1 nodes 151865736 nps 82133983 pv f1 b1 g2 c1 g3 c1e1
+info depth 7 time 62967 score 3 nodes 5322126040 nps 84522464 pv f2 a2 e1 f6 d1 a2a4 d1b1
+```
+
+```
+1+0.01:
+Score of AtaxxDotCpp-dev vs AtaxxDotCpp-master: 5000 - 0 - 0  [1.000] 5000
+inf +/- -nan
 ```